split_still_data: prune shared stills imageset to surviving experiments
dials.split_still_data splits via save_subset (util/image_grouping.py),
which filters experiments with select_on_experiment_identifiers then
writes with as_file -- but never pruned the shared stills ImageSequence.
The survivors kept pointing at an imageset advertising all source frames
(single_file_indices) while only the selected experiments contributed
scan_points, so to_dict consolidation wrote fewer scan_points than the
imageset's frames, tripping the read-side _expand_consolidated_scans
assert len(single_file_indices) == unique scan_points
(test_split_still_data.py::test_split_still_data_h5: 875 != 15).

Structurally identical to the combine_experiments min/max prune bug
(f4e5155c9 / cfd2cb741). Reuse that exact prune rather than reimplement:
its len(iset) <= 1 guard (keeps multi-file CBF from spurious rebuild),
reuse-data()-without-reopen branch (required -- save_subset loads
check_format=False, so re-opening source would fail), and imageset_id
remap are all needed.

Move _consolidate_stills_imagesets from command_line/combine_experiments.py
into the existing util/combine_experiments.py as public
consolidate_stills_imagesets (fully self-contained; avoids a util ->
command_line dependency inversion / circular-import risk);
command_line/combine_experiments.py imports it back. save_subset calls it
on the finalized expts/refls just before as_file (no-op for non-stills /
per-frame imagesets).

Maintains the P2 single_file_indices == unique scan_points invariant for
the split path; no to_dict change (the systemic fix remains re-scoped P3).
Verified: test_split_still_data.py (3/3, incl. cunir multi-file CBF) +
test_combine_experiments.py (18/18) green.
